requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.7.0
APScheduler==3.10.4
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
multidict==6.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-telegram-bot==20.0a6
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
yarl==1.18.3
